Corrigir erros nos vectors

+++ include/math/vectors.inl
@@ -1,7 +1,7 @@
 #include "../vector3.h"
 
 inline Vector3 cross(Vector3 u, Vector3 v) {
-    int uvi, uvj, uvk;
+    double uvi, uvj, uvk;
     uvi = (u[1] * v[2]) - (u[2] * v[1]);
     uvj = (u[2] * v[0]) - (u[0] * v[2]);
     uvk = (u[0] * v[1]) - (u[1] * v[0]);
